revert some test code
